augur-engine/requirements.txt
httpx>=0.27
anthropic>=0.39
feedparser>=6.0
Pillow>=10.0
python-dotenv>=1.0
python-dateutil>=2.8
